+++ src/include/tclap/MultiArg.h
@@ -165,7 +165,7 @@ public:
      * \param i - Pointer the the current argument in the list.
      * \param args - Mutable list of strings. Passed from main().
      */
-    virtual bool processArg(int *i, std::vector<std::string> &args);
+    virtual bool processArg(size_t *i, std::vector<std::string> &args);
 
     /**
      * Returns a vector of type T containing the values parsed from
@@ -265,7 +265,7 @@ MultiArg<T>::MultiArg(const std::string &flag, const std::string &name,
 }
 
 template <class T>
-bool MultiArg<T>::processArg(int *i, std::vector<std::string> &args) {
+bool MultiArg<T>::processArg(size_t *i, std::vector<std::string> &args) {
     if (_hasBlanks(args[*i])) return false;
 
     std::string flag = args[*i];
